Namespaces: separate tests from source
